Abstracting RCC implementations
